# perf(Rust): Use SIMD to se/de string