--- a/rust/fory-core/benches/simd_bench.rs
+++ b/rust/fory-core/benches/simd_bench.rs
@@ -19,6 +19,12 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 #[cfg(target_feature = "avx2")]
 use std::arch::x86_64::*;
 
+use fory_core::buffer::{Reader, Writer};
+use fory_core::meta::buffer_rw_string::{
+    read_latin1_simd, read_latin1_standard, read_utf16_simd, read_utf16_standard, read_utf8_simd,
+    read_utf8_standard, write_latin1_simd, write_latin1_standard, write_utf16_simd,
+    write_utf16_standard, write_utf8_simd, write_utf8_standard,
+};
 #[cfg(target_feature = "sse2")]
 use std::arch::x86_64::*;
 
@@ -78,6 +84,168 @@ fn is_latin_std(s: &str) -> bool {
     s.bytes().all(|b| b.is_ascii())
 }
 
+fn benchmark_write_latin1(c: &mut Criterion) {
+    let sizes = [100, 1000, 10000, 100000];
+    let ascii_string = "abcdefghijklmnopqrstuvwxyz0123456789";
+
+    for &size in &sizes {
+        let s_ascii = ascii_string.repeat(size / ascii_string.len() + 1);
+
+        let name_simd = format!("Latin-1 SIMD size {}", size);
+        c.bench_function(&name_simd, |b| {
+            b.iter(|| {
+                let mut w = Writer::default();
+                write_latin1_simd(black_box(&mut w), black_box(&s_ascii));
+            })
+        });
+
+        let name_scalar = format!("Latin-1 Standard size {}", size);
+        c.bench_function(&name_scalar, |b| {
+            b.iter(|| {
+                let mut w = Writer::default();
+                write_latin1_standard(black_box(&mut w), black_box(&s_ascii));
+            })
+        });
+    }
+}
+
+fn benchmark_write_utf8(c: &mut Criterion) {
+    let sizes = [100, 1000, 10000, 100000];
+    for &size in &sizes {
+        let s = "Hello, 世界! 🌍".repeat(size);
+
+        let name_simd = format!("UTF-8 SIMD size {}", size);
+        c.bench_function(&name_simd, |b| {
+            b.iter(|| {
+                let mut w = Writer::default();
+                write_utf8_simd(black_box(&mut w), black_box(&s));
+            })
+        });
+
+        let name_scalar = format!("UTF-8 Standard size {}", size);
+        c.bench_function(&name_scalar, |b| {
+            b.iter(|| {
+                let mut w = Writer::default();
+                write_utf8_standard(black_box(&mut w), black_box(&s));
+            })
+        });
+    }
+}
+
+fn benchmark_write_utf16(c: &mut Criterion) {
+    let sizes = [100, 1000, 10000, 100000];
+    for &size in &sizes {
+        let s = "Hello, 世界! 🌍".repeat(size);
+
+        let name_simd = format!("UTF-16 SIMD size {}", size);
+        c.bench_function(&name_simd, |b| {
+            b.iter(|| {
+                let mut w = Writer::default();
+                let utf16: Vec<u16> = s.encode_utf16().collect();
+                write_utf16_simd(black_box(&mut w), black_box(&utf16));
+            })
+        });
+
+        let name_scalar = format!("UTF-16 Standard size {}", size);
+        c.bench_function(&name_scalar, |b| {
+            b.iter(|| {
+                let mut w = Writer::default();
+                let utf16: Vec<u16> = s.encode_utf16().collect();
+                write_utf16_standard(black_box(&mut w), black_box(&utf16));
+            })
+        });
+    }
+}
+
+fn benchmark_read_latin1(c: &mut Criterion) {
+    let sizes = [100, 1000, 10000, 100000];
+    let ascii_string = "abcdefghijklmnopqrstuvwxyz0123456789";
+
+    for &size in &sizes {
+        let s_ascii = ascii_string.repeat(size / ascii_string.len() + 1);
+        let mut writer = Writer::default();
+        writer.write_latin1_string(&s_ascii);
+        let data = writer.dump();
+
+        let name_simd = format!("Read Latin-1 SIMD size {}", size);
+        c.bench_function(&name_simd, |b| {
+            b.iter(|| {
+                let mut reader = Reader::new(black_box(&data));
+                read_latin1_simd(black_box(&mut reader), black_box(s_ascii.len()));
+            })
+        });
+
+        let name_scalar = format!("Read Latin-1 Standard size {}", size);
+        c.bench_function(&name_scalar, |b| {
+            b.iter(|| {
+                let mut reader = Reader::new(black_box(&data));
+                read_latin1_standard(black_box(&mut reader), black_box(s_ascii.len()));
+            })
+        });
+    }
+}
+
+fn benchmark_read_utf8(c: &mut Criterion) {
+    let sizes = [100, 1000, 10000, 100000];
+    let test_string = "Hello, 世界! 🌍";
+
+    for &size in &sizes {
+        let s = test_string.repeat(size / test_string.len() + 1);
+        let mut writer = Writer::default();
+        writer.write_utf8_string(&s);
+        let data = writer.dump();
+
+        let name_simd = format!("Read UTF-8 SIMD size {}", size);
+        c.bench_function(&name_simd, |b| {
+            b.iter(|| {
+                let mut reader = Reader::new(black_box(&data));
+                read_utf8_simd(black_box(&mut reader), black_box(s.len()));
+            })
+        });
+
+        let name_scalar = format!("Read UTF-8 Standard size {}", size);
+        c.bench_function(&name_scalar, |b| {
+            b.iter(|| {
+                let mut reader = Reader::new(black_box(&data));
+                read_utf8_standard(black_box(&mut reader), black_box(s.len()));
+            })
+        });
+    }
+}
+
+fn benchmark_read_utf16(c: &mut Criterion) {
+    let sizes = [100, 1000, 10000, 100000];
+    let test_string = "Hello, 世界! 🌍";
+
+    for &size in &sizes {
+        let s = test_string.repeat(size / test_string.len() + 1);
+        // 构造 UTF-16 LE 字节数据
+        let mut data: Vec<u8> = Vec::with_capacity(s.len() * 2);
+        for u in s.encode_utf16() {
+            let lo = (u & 0x00FF) as u8;
+            let hi = (u >> 8) as u8;
+            data.push(lo);
+            data.push(hi);
+        }
+
+        let name_simd = format!("Read UTF-16 SIMD size {}", size);
+        c.bench_function(&name_simd, |b| {
+            b.iter(|| {
+                let mut reader = Reader::new(black_box(&data));
+                read_utf16_simd(black_box(&mut reader), black_box(data.len()));
+            })
+        });
+
+        let name_scalar = format!("Read UTF-16 Standard size {}", size);
+        c.bench_function(&name_scalar, |b| {
+            b.iter(|| {
+                let mut reader = Reader::new(black_box(&data));
+                read_utf16_standard(black_box(&mut reader), black_box(data.len()));
+            })
+        });
+    }
+}
+
 fn criterion_benchmark(c: &mut Criterion) {
     let test_str_short = "Hello, World!";
     let test_str_long = "Hello, World! ".repeat(1000);
@@ -106,6 +274,14 @@ fn criterion_benchmark(c: &mut Criterion) {
     c.bench_function("Standard long", |b| {
         b.iter(|| is_latin_std(black_box(&test_str_long)))
     });
+
+    benchmark_write_latin1(c);
+    benchmark_write_utf8(c);
+    benchmark_write_utf16(c);
+
+    benchmark_read_latin1(c);
+    benchmark_read_utf8(c);
+    benchmark_read_utf16(c);
 }
 
 criterion_group!(benches, criterion_benchmark);