--- a/rust/fory-core/src/meta/mod.rs
+++ b/rust/fory-core/src/meta/mod.rs
@@ -24,7 +24,7 @@ pub use meta_string::{
     FIELD_NAME_DECODER, FIELD_NAME_ENCODER, NAMESPACE_DECODER, NAMESPACE_ENCODER,
     TYPE_NAME_DECODER, TYPE_NAME_ENCODER,
 };
-pub use string_util::{get_latin1_length, is_latin, murmurhash3_x64_128};
+pub use string_util::{buffer_rw_string, get_latin1_length, is_latin, murmurhash3_x64_128};
 pub use type_meta::{
     FieldInfo, FieldType, NullableFieldType, TypeMeta, TypeMetaLayer, NAMESPACE_ENCODINGS,
     TYPE_NAME_ENCODINGS,